clean: Extract re-usable ObjectMapper utility from JacksonObjectReaderWriter
