+++ java/dev/enola/common/io/object/jackson/JacksonObjectReaderWriter.java
@@ -17,17 +17,9 @@
  */
 package dev.enola.common.io.object.jackson;
 
-import com.fasterxml.jackson.annotation.JsonInclude;
-import com.fasterxml.jackson.annotation.JsonSetter;
-import com.fasterxml.jackson.annotation.Nulls;
-import com.fasterxml.jackson.core.json.JsonReadFeature;
 import com.fasterxml.jackson.databind.MappingIterator;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.cfg.CoercionAction;
-import com.fasterxml.jackson.databind.cfg.CoercionInputShape;
 import com.fasterxml.jackson.databind.module.SimpleModule;
-import com.fasterxml.jackson.databind.type.LogicalType;
-import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
 import com.google.common.net.MediaType;
 import com.google.errorprone.annotations.CanIgnoreReturnValue;
 
@@ -36,14 +28,13 @@ import dev.enola.common.io.object.ObjectReaderWriter;
 import dev.enola.common.io.object.ProviderFromID;
 import dev.enola.common.io.resource.ReadableResource;
 import dev.enola.common.io.resource.WritableResource;
+import dev.enola.common.jackson.ObjectMappers;
 
 import org.jspecify.annotations.Nullable;
 
 import java.io.IOException;
 import java.util.List;
-import java.util.Map;
 import java.util.Optional;
-import java.util.Set;
 
 abstract class JacksonObjectReaderWriter implements ObjectReaderWriter {
 
@@ -65,45 +56,12 @@ abstract class JacksonObjectReaderWriter implements ObjectReaderWriter {
     }
 
     private void init() {
-        // Do NOT use mapper.findAndRegisterModules();
-        // because that would mean that mapping would depend on (uncontrollable) classpath
-
         var module = new SimpleModule();
         module.addSerializer(Identifiable.class, new IdentifiableIdSerializer());
         if (provider != null) module.setDeserializers(new IdentifiableDeserializers(provider));
         mapper.registerModule(module);
 
-        // https://github.com/FasterXML/jackson-modules-java8
-        mapper.registerModule(new JavaTimeModule());
-        // TODO Optional<T> support with mapper.registerModule(new Jdk8Module());
-
-        // Enable JSONc features: Java-style comments and trailing commas
-        // https://github.com/enola-dev/enola/issues/1847
-        mapper.getFactory().enable(JsonReadFeature.ALLOW_JAVA_COMMENTS.mappedFeature());
-        mapper.getFactory().enable(JsonReadFeature.ALLOW_TRAILING_COMMA.mappedFeature());
-
-        // DO fail on unknown properties - this helps to spot errors in configuration files etc.
-        // NO! mapper.configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false);
-
-        // Always skip empty sequences ([]) and maps
-        mapper.setSerializationInclusion(JsonInclude.Include.NON_EMPTY);
-
-        // Always set missing sequences to empty collections
-        var nullAsEmpty = JsonSetter.Value.forValueNulls(Nulls.AS_EMPTY);
-        mapper.configOverride(List.class).setSetterInfo(nullAsEmpty);
-        mapper.configOverride(Set.class).setSetterInfo(nullAsEmpty);
-        mapper.configOverride(Map.class).setSetterInfo(nullAsEmpty);
-
-        // Always allow coercion for e.g. empty Map keys etc.
-        mapper.coercionConfigFor(LogicalType.POJO)
-                .setAcceptBlankAsEmpty(true)
-                .setCoercion(CoercionInputShape.EmptyString, CoercionAction.AsNull);
-        mapper.coercionConfigFor(LogicalType.Array)
-                .setAcceptBlankAsEmpty(true)
-                .setCoercion(CoercionInputShape.EmptyString, CoercionAction.AsNull);
-        mapper.coercionConfigFor(LogicalType.Map)
-                .setAcceptBlankAsEmpty(true)
-                .setCoercion(CoercionInputShape.EmptyString, CoercionAction.AsNull);
+        ObjectMappers.configure(mapper);
     }
 
     abstract boolean canHandle(MediaType mediaType);
