+++ java/dev/enola/common/jackson/package-info.java
@@ -1,0 +1,29 @@
+/*
+ * SPDX-License-Identifier: Apache-2.0
+ *
+ * Copyright 2025 The Enola <https://enola.dev> Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+/**
+ * Jackson utilities.
+ *
+ * <p>See also <code>package dev.enola.common.io.object.jackson</code>; this package is for more
+ * general purpose Jackson utilities that do not depend on Enola's <code>dev.enola.common.io.object
+ * </code> framework, making them usable in other contexts as well.
+ */
+@NullMarked
+package dev.enola.common.jackson;
+
+import org.jspecify.annotations.NullMarked;
